refactor(frame_server): split module into transport, push, and command mixins
frame_server.py (2,100+ lines) becomes the frame_server/ package:
- frames.py: shared contact/advert frame-body encoders
- transport.py: server lifecycle, frame framing, writer task, client loop
- push.py: bridge event subscriptions (inline closures now named methods)
  and public push_* methods
- commands_{contacts,channels,messaging,device}.py: _cmd_* handlers by domain
- server.py: CompanionFrameServer class, dispatch table, persistence hooks

All methods remain on CompanionFrameServer via mixins and the import path
openhop_core.companion.frame_server.CompanionFrameServer is unchanged, so
downstream subclasses (openhop_repeater) are unaffected.
